requirements.txt
anyio==4.8.0
confluent-kafka==2.8.0
fastapi==0.115.6
grpcio==1.70.0
grpcio-reflection==1.70.0
httpx==0.28.1
locust==2.24.0
motor==3.6.1
openai==1.65.4
prometheus_client==0.21.1
prometheus-fastapi-instrumentator==7.0.2
prometheus-kafka-metrics==0.4.2 
protobuf==5.29.3
pydantic==2.10.5
py-grpc-prometheus==0.8.0
pymongo==4.9.2
pyopenssl==24.0.0
requests==2.32.3
starlette==0.41.3
uvicorn==0.34.0